Moved code to extern files
