+++ templates/template.cs
@@ -1,0 +1,11 @@
+namespace PLACEHOLDER
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hi mom");
+            Console.ReadLine();
+        }
+    }
+}
